CurrentUrlProviderを独立したファイルに移動してreact-refreshのエラーを抑制
