+++ src/main.tsx
@@ -1,18 +1,18 @@
 /// <reference types="react/experimental" />
 
-import React from "react";
-import ReactDOM from "react-dom/client";
 import { HeroUIProvider } from "@heroui/react";
 import { ThemeProvider } from "next-themes";
-import App from "./App.tsx";
-import "./index.css";
-import CacheProvider from "./CacheProvider.tsx";
+import React from "react";
+import ReactDOM from "react-dom/client";
 import { SWRConfig } from "swr";
+import App from "./App.tsx";
+import CacheProvider from "./CacheProvider.tsx";
 import {
   ChromeLocalStorageProvider,
   ChromeSessionStorageProvider,
 } from "./contexts/ChromeStorageContext.tsx";
-import { CurrentUrlProvider } from "./contexts/CurrentUrlContext.tsx";
+import CurrentUrlProvider from "./CurrentUrlProvider.tsx";
+import "./index.css";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
